Colourize the switch table

+++ include/tasks/mission_control.h
@@ -7,7 +7,6 @@
 
 void mission_control(void) __attribute__((noreturn));
 
-int  update_turnout(short num, short state);
+int update_turnout(int num, int state);
 
 #endif
-
